requirements.txt
verboselogs
coloredlogs
discord.py
python-dotenv